feat: move cloud-provider and cloud-region subcommands as root commands
It makes more sense this way, as those are not dependant on clusters
existing
